Change balance to float

+++ app/models/user.py
@@ -3,4 +3,4 @@ from pydantic import Field, BaseModel
 
 class User(BaseModel):
     id: str = Field(frozen=True)
-    balance: int = Field(ge=0)
+    balance: float = Field(ge=0)
